requirements.txt
beautifulsoup4==4.9.3
html5lib==1.0.1
lxml==4.6.2
requests==2.22.0
selenium==3.141.0
python-dotenv==0.15.0
python-telegram-bot==13.4.1

